requirements.txt
textual>=0.47.0
rich>=13.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0
